use import checks, single-gpu mode can still be used without installing Horovod
