+++ athena/solver.py
@@ -21,6 +21,10 @@ import warnings
 import time
 import tensorflow as tf
 from absl import logging
+try:
+    import horovod.tensorflow as hvd
+except ImportError:
+    print("There is some problem with your horovod installation. But it wouldn't affect single-gpu training")
 from .utils.hparam import register_and_parse_hparams
 from .utils.metric_check import MetricChecker
 from .utils.misc import validate_seqs
@@ -115,6 +119,56 @@ class BaseSolver(tf.keras.Model):
         self.model.reset_metrics()
         return loss_metric.result()
 
+
+class HorovodSolver(BaseSolver):
+    """ A multi-processer solver based on Horovod """
+
+    @staticmethod
+    def initialize_devices(visible_gpu_idx=None):
+        """ initialize hvd devices, should be called firstly """
+        if visible_gpu_idx is not None:
+            warnings.warn("we can not set the visible gpu idx like this")
+        hvd.init()
+        gpus = tf.config.experimental.list_physical_devices("GPU")
+        for gpu in gpus:
+            tf.config.experimental.set_memory_growth(gpu, True)
+        if gpus:
+            tf.config.experimental.set_visible_devices(gpus[hvd.local_rank()], "GPU")
+
+    def train_step(self, samples):
+        """ train the model 1 step """
+        with tf.GradientTape() as tape:
+            logits = self.model(samples, training=True)
+            loss, metrics = self.model.get_loss(logits, samples, training=True)
+        # Horovod: add Horovod Distributed GradientTape.
+        tape = hvd.DistributedGradientTape(tape)
+        grads = tape.gradient(loss, self.model.trainable_variables)
+        grads = self.clip_by_norm(grads, self.hparams.clip_norm)
+        self.optimizer.apply_gradients(zip(grads, self.model.trainable_variables))
+        return loss, metrics
+
+    def train(self, dataset, total_batches=-1):
+        """ Update the model in 1 epoch """
+        train_step = self.train_step
+        if self.hparams.enable_tf_function:
+            logging.info("please be patient, enable tf.function, it takes time ...")
+            train_step = tf.function(train_step, input_signature=self.sample_signature)
+        for batch, samples in enumerate(dataset.take(total_batches)):
+            # train 1 step
+            samples = self.model.prepare_samples(samples)
+            loss, metrics = train_step(samples)
+            # Horovod: broadcast initial variable states from rank 0 to all other processes.
+            # This is necessary to ensure consistent initialization of all workers when
+            # training is started with random weights or restored from a checkpoint.
+            #
+            # Note: broadcast should be done after the first gradient step to ensure optimizer
+            # initialization.
+            if batch == 0:
+                hvd.broadcast_variables(self.model.trainable_variables, root_rank=0)
+                hvd.broadcast_variables(self.optimizer.variables(), root_rank=0)
+            if batch % self.hparams.log_interval == 0 and hvd.local_rank() == 0:
+                logging.info(self.metric_checker(loss, metrics))
+                self.model.reset_metrics()
 
 
 class DecoderSolver(BaseSolver):
